worker-python/requirements.txt
# Python AI Worker Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9

# Existing extraction tools
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
pyyaml==6.0.1
lxml==5.1.0
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0

